+++ jest.config.js
@@ -1,4 +1,4 @@
 module.exports = {
   testEnvironment: 'jsdom',
-  setupFiles: ['<rootDir>/tests/setup.js'],
+  setupFiles: ['<rootDir>/tests/setup.js']
 };